+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 
 ### New
 
+- [#838](https://gitlab.com/meltano/meltano/issues/838) Add indicator for speed run plugins
 - [#870](https://gitlab.com/meltano/meltano/issues/870) Add global footer component in docs
 - [#871](https://gitlab.com/meltano/meltano/issues/871) Add contributing link in footer of docs
 - [#908](https://gitlab.com/meltano/meltano/issues/908) Add auto installation for Airflow Orchestrator for improved UX
@@ -18,8 +19,11 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 
 ### Changes
 
+- [#838](https://gitlab.com/meltano/meltano/issues/838) Speed run plugins prioritized to top of the list
 - [#896](https://gitlab.com/meltano/meltano/issues/896) Add documentation for how to do patch releases
 - [#910](https://gitlab.com/meltano/meltano/issues/910) Update linting rules to enforce better standards for the frontend code base
+- [#885](https://gitlab.com/meltano/meltano/issues/885) Add docs for all extractors and loaders
+- [#885](https://gitlab.com/meltano/meltano/issues/885) All plugin modal cards show docs text if they have docs
 
 ### Fixes